requirements.txt
aiogram
peewee
loguru
python-dotenv
